Fix bug, split quilt
